Convert chronological data to an sf object

chron_as_sf() adds a simple features geometry column (see sf::st_sf) to chronological data from XRONOS, using the latitude and longitude columns.

chron_drop_na_coords() excludes rows with missing or invalid coordinates; it is implicitly called by chron_as_sf(), issuing a warning if any rows were removed.

chron_as_sf(x, crs = sf::st_crs(4326))

chron_drop_na_coords(x, .warn = FALSE)

Arguments

x
data.frame retrieved with chron_data().

crs
Desired coordinate reference system in a format understood by sf::st_crs(). Defaults to latitude/longitude on the WGS84 ellipsoid (EPSG:4326).

.warn
If TRUE, issues a warning when rows are removed.

Value

chron_as_sf() returns x converted into an sf object. The original coordinate columns are retained.

chron_drop_na_coords() returns x without rows with missing or invalid coordinates.

Examples

x <- chron_data(country = "Switzerland")
chron_as_sf(x)

#> Warning: 18 rows with missing or invalid coordinates were removed.

#> Simple feature collection with 1161 features and 21 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 6.1418 ymin: 46.0211 xmax: 10.3818 ymax: 47.75
#> Geodetic CRS:  WGS 84
#> # A tibble: 1,161 × 22
#>       id labnr    reference source_database    bp   std cal_bp cal_std delta_c13
#>  * <int> <chr>    <lgl>     <chr>           <int> <int>  <int>   <int>     <dbl>
#>  1  7878 ETH-111… NA        calpal           3845    60   4261     164         0
#>  2  7879 UZ-951   NA        calpal           5020    90   5767     164         0
#>  3  7880 ETH-261… NA        calpal           5050    60   5801     135         0
#>  4  7881 KN-121   NA        calpal           5080    70   5814     152         0
#>  5  7884 ETH-5597 NA        calpal           8175   100   9142     312         0
#>  6  7885 ETH-9570 NA        calpal           7745    85   8527     204         0
#>  7  7886 ETH-226… NA        calpal           4110    55   4641     180         0
#>  8  7887 ETH-976… NA        calpal           4905    65   5643     188         0
#>  9  7888 ETH-976… NA        calpal           4910    65   5648     188         0
#> 10  7889 ETH-976… NA        calpal           4825    70   5540     190         0
#> # … with 1,151 more rows, and 13 more variables: site <chr>, site_phase <chr>,
#> #   feature <chr>, period <lgl>, typochronological_unit <lgl>,
#> #   ecochronological_unit <lgl>, material <chr>, lat <chr>, lng <chr>,
#> #   country <chr>, species <chr>, site_type <chr>, geometry <POINT [°]>