Get chronological data from XRONOS

Retrieves chronological data from XRONOS. chron_data() retrieves all records. Additional parameters can be used to filter the results.

chron_data(..., .everything = NA)

Arguments

...

(Optional) One or more named arguments specifying filter variables and values to include. Supported filters are "labnr", "site", "site_type", "country", "feature", "material", and "species".

.everything

If TRUE, suppresses interactive mode prompt when retrieving all records from XRONOS (see details).

Details

To reduce unnecessary server load, get_xronos() (without any filters) will prompt for confirmation in interactive mode interactive mode. Set .everything = TRUE to suppress this.

Examples

# Dates on charcoal or bone from Switzerland
chron_data(country = "Switzerland", material = c("charcoal", "bone"))
#> # A tibble: 558 × 21
#>       id labnr  refer…¹ sourc…²    bp   std cal_bp cal_std delta…³ site  site_…⁴
#>    <int> <chr>  <lgl>   <chr>   <int> <int>  <int>   <int>   <dbl> <chr> <chr>  
#>  1 24781 ETH-5… NA      radonb   3355    30   3582      98      NA Val … Val Ur…
#>  2 24783 ETH-3… NA      radonb   2190    50   2205     148      NA Val … Val Ta…
#>  3 24800 ETH-3… NA      radonb   2650    35   2762      78      NA Val … Val Ur…
#>  4 24807 ETH-5… NA      radonb   2567    30   2722     189      NA Val … Val Ur…
#>  5 27588 ETH-3… NA      radonb   3060    40   3272     103      NA Schn… Schnap…
#>  6 27595 ETH-6… NA      radonb   3467    29   3745      96      NA Val … Val Fe…
#>  7 27599 ETH-5… NA      radonb   2882    29   3010     122      NA Val … Val Ta…
#>  8 27600 ETH-6… NA      radonb   2518    25   2585     135      NA Val … Val Ta…
#>  9 27659 ETH-5… NA      radonb   3356    30   3584      98      NA Val … Val Ur…
#> 10 27746 ETH-4… NA      radonb   2520    25   2586     135      NA Val … Val Ur…
#> # … with 548 more rows, 10 more variables: site_type <chr>, feature <chr>,
#> #   periods <list>, typochronological_units <list>,
#> #   ecochronological_units <list>, material <chr>, lat <chr>, lng <chr>,
#> #   country <chr>, species <chr>, and abbreviated variable names ¹​reference,
#> #   ²​source_database, ³​delta_c13, ⁴​site_phase