Convert chronological data to an sf object

chron_as_sf() adds a simple features geometry column (see sf::st_sf) to chronological data from XRONOS, using the latitude and longitude columns.

chron_drop_na_coords() excludes rows with missing or invalid coordinates; it is implicitly called by chron_as_sf(), issuing a warning if any rows were removed.

chron_as_sf(x, crs = sf::st_crs(4326))

chron_drop_na_coords(x, .warn = FALSE)

Arguments

x

data.frame retrieved with chron_data().

crs

Desired coordinate reference system in a format understood by sf::st_crs(). Defaults to latitude/longitude on the WGS84 ellipsoid (EPSG:4326).

.warn

If TRUE, issues a warning when rows are removed.

Value

chron_as_sf() returns x converted into an sf object. The original coordinate columns are retained. chron_drop_na_coords() returns x without rows with missing or invalid coordinates.

Examples

x <- chron_data(country = "Switzerland")
chron_as_sf(x)
#> Warning: 16 rows with missing or invalid coordinates were removed.
#> Simple feature collection with 1252 features and 21 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 4.0911 ymin: 46.0211 xmax: 10.3818 ymax: 47.753
#> Geodetic CRS:  WGS 84
#> # A tibble: 1,252 × 22
#>       id labnr    reference source_database    bp   std cal_bp cal_std delta_c13
#>  * <int> <chr>    <lgl>     <chr>           <int> <int>  <int>   <int>     <dbl>
#>  1 86751 ETH-260… NA        calpal           3130    50   3346     119         0
#>  2 86754 ETH-260… NA        calpal           3215    50   3427     123         0
#>  3 86761 ETH-260… NA        calpal           3205    50   3420     129         0
#>  4 86768 ETH-274… NA        calpal           3290    50   3511     121         0
#>  5 86788 ETH-289… NA        calpal           3455    50   3718     138         0
#>  6 86812 ETH-290… NA        calpal           3420    50   3666     155         0
#>  7 86819 ETH-290… NA        calpal           3255    55   3471     121         0
#>  8 86821 ETH-290… NA        calpal           3240    55   3454     115         0
#>  9 86828 ETH-336… NA        calpal           2935    55   3088     178         0
#> 10 86841 UZ-4715… NA        calpal           3185    50   3404     132         0
#> # … with 1,242 more rows, and 13 more variables: site <chr>, site_phase <chr>,
#> #   feature <chr>, period <named list>, typochronological_unit <named list>,
#> #   ecochronological_unit <named list>, material <chr>, lat <chr>, lng <chr>,
#> #   country <chr>, site_type <chr>, species <chr>, geometry <POINT [°]>