Get chronological data from XRONOS

Retrieves chronological data from XRONOS. chron_data() retrieves all records. Additional parameters can be used to filter the results.

chron_data(..., .everything = NA)

Arguments

...

(Optional) One or more named arguments specifying filter variables and values to include. Supported filters are "labnr", "site", "site_type", "country", "feature", "material", and "species".

.everything

If TRUE, suppresses interactive mode prompt when retrieving all records from XRONOS (see details).

Details

To reduce unnecessary server load, get_xronos() (without any filters) will prompt for confirmation in interactive mode interactive mode. Set .everything = TRUE to suppress this.

Examples

# Dates on charcoal or bone from Switzerland
chron_data(country = "Switzerland", material = c("charcoal", "bone"))
#> # A tibble: 558 × 21
#>       id labnr    reference source_database    bp   std cal_bp cal_std delta_c13
#>    <int> <chr>    <lgl>     <chr>           <int> <int>  <int>   <int>     <dbl>
#>  1 24781 ETH-599… NA        radonb           3355    30   3582      98        NA
#>  2 86489 ETH-9647 NA        calpal           3885    80   4303     219         0
#>  3 86500 ETH-111… NA        calpal           3845    60   4258     166         0
#>  4 86521 ETH-9653 NA        calpal           4490    75   5134     243         0
#>  5 86561 B-3395   NA        calpal           3835    60   4243     172         0
#>  6 86564 B-3610   NA        calpal           2960    70   3122     208         0
#>  7 86572 Ua-20702 NA        calpal           2610    45   2741     194         0
#>  8 86573 ETH-284… NA        calpal           3505    50   3773     137         0
#>  9 86579 ETH-331… NA        calpal           3305    50   3525     132         0
#> 10 86581 ETH-331… NA        calpal           3285    50   3505     121         0
#> # … with 548 more rows, and 12 more variables: site <chr>, site_phase <chr>,
#> #   site_type <chr>, feature <chr>, period <named list>,
#> #   typochronological_unit <named list>, ecochronological_unit <named list>,
#> #   material <chr>, lat <chr>, lng <chr>, country <chr>, species <chr>