Get chronological data from XRONOS

Retrieves chronological data from XRONOS. chron_data() retrieves all records. Additional parameters can be used to filter the results.

chron_data(..., .everything = NA)

Arguments

...

(Optional) One or more named arguments specifying filter variables and values to include. Supported filters are "labnr", "site", "site_type", "country", "feature", "material", and "species".

.everything

If TRUE, suppresses interactive mode prompt when retrieving all records from XRONOS (see details).

Details

To reduce unnecessary server load, get_xronos() (without any filters) will prompt for confirmation in interactive mode interactive mode. Set .everything = TRUE to suppress this.

country should be a valid ISO two-letter country code. Otherwise, the function will try to interpret as such using countrycode::countryname(), or return an error if any values of country cannot be matched.

Examples

# Dates on charcoal or bone from Switzerland
chron_data(country = "Switzerland", material = c("charcoal", "bone"))
#> # A tibble: 1,847 × 21
#>       id labnr      bp   std cal_bp cal_std delta…¹ reference    sourc…² mater…³
#>    <int> <chr>   <int> <int> <lgl>  <lgl>     <dbl> <list>       <chr>   <chr>  
#>  1  4946 ETH-63…  9044    35 NA     NA           NA <named list> AgriCh… charco…
#>  2  4947 ETH-63…  8912    33 NA     NA           NA <named list> AgriCh… charco…
#>  3  4948 ETH-63…  8909    47 NA     NA           NA <named list> AgriCh… charco…
#>  4  4949 ETH-63…  8812    33 NA     NA           NA <named list> AgriCh… charco…
#>  5  5070 VERA-2…   235    35 NA     NA           NA <named list> AgriCh… charco…
#>  6  5134 UZ-2480  5020    75 NA     NA           NA <named list> AgriCh… charco…
#>  7  5155 CRG-582  5950   100 NA     NA           NA <named list> AgriCh… charco…
#>  8  5156 CRG-580  5980   175 NA     NA           NA <named list> AgriCh… charco…
#>  9  5157 CRG-581  5715   160 NA     NA           NA <named list> AgriCh… charco…
#> 10  5158 CRG-430  5484   128 NA     NA           NA <named list> AgriCh… charco…
#> # … with 1,837 more rows, 11 more variables: species <chr>, feature <chr>,
#> #   periods <list>, typochronological_units <list>,
#> #   ecochronological_units <list>, site <chr>, country <chr>, lat <chr>,
#> #   lng <chr>, site_type <chr>, feature_type <chr>, and abbreviated variable
#> #   names ¹​delta_c13, ²​source_database, ³​material