Get chronological data from XRONOS

Retrieves chronological data from XRONOS. chron_data() retrieves all records. Additional parameters can be used to filter the results.

chron_data(..., .everything = NA)

Arguments

...

(Optional) One or more named arguments specifying filter variables and values to include. Supported filters are "labnr", "site", "site_type", "country", "feature", "material", and "species".

.everything

If TRUE, suppresses interactive mode prompt when retrieving all records from XRONOS (see details).

Value

A data frame with the returned data from XRONOS.

Details

To reduce unnecessary server load, get_xronos() (without any filters) will prompt for confirmation in interactive mode interactive mode. Set .everything = TRUE to suppress this.

country should be a valid ISO two-letter country code. Otherwise, the function will try to interpret as such using countrycode::countryname(), or return an error if any values of country cannot be matched.

Examples

# Dates on charcoal or bone from Switzerland
chron_data(country = "Switzerland", material = c("charcoal", "bone"))
#> # A tibble: 3,320 × 20
#>       id labnr        bp   std source_database lab_name material species        
#>    <dbl> <chr>     <dbl> <dbl> <chr>           <chr>    <chr>    <chr>          
#>  1  4946 ETH-63150  9044    35 ""              ""       charcoal Corylus avella…
#>  2  4946 ETH-63150  9044    35 ""              ""       charcoal Corylus avella…
#>  3  4947 ETH-63685  8912    33 ""              ""       charcoal Corylus avella…
#>  4  4947 ETH-63685  8912    33 ""              ""       charcoal Corylus avella…
#>  5  4948 ETH-63687  8909    47 ""              ""       charcoal Ulmus          
#>  6  4948 ETH-63687  8909    47 ""              ""       charcoal Ulmus          
#>  7  4949 ETH-63686  8812    33 ""              ""       charcoal Prunus         
#>  8  4949 ETH-63686  8812    33 ""              ""       charcoal Prunus         
#>  9  5070 VERA-2905   235    35 ""              ""       charcoal NA             
#> 10  5070 VERA-2905   235    35 ""              ""       charcoal NA             
#> # ℹ 3,310 more rows
#> # ℹ 12 more variables: feature_type <chr>, site <chr>, country <chr>,
#> #   site_type <chr>, periods <list>, typochronological_units <list>,
#> #   ecochronological_units <list>, reference <list>, lat <chr>, lng <chr>,
#> #   delta_c13 <dbl>, feature <chr>