Get chronological data from XRONOS

Retrieves chronological data from XRONOS. chron_data() retrieves all records. Additional parameters can be used to filter the results.

chron_data(..., .everything = NA)

Arguments

...
(Optional) One or more named arguments specifying filter variables and values to include. Supported filters are "labnr", "site", "site_type", "country", "feature", "material", and "species".

.everything
If TRUE, suppresses interactive mode prompt when retrieving all records from XRONOS (see details).

Details

To reduce unnecessary server load, get_xronos() (without any filters) will prompt for confirmation in interactive mode interactive mode. Set .everything = TRUE to suppress this.

Examples

# Dates on charcoal or bone from Switzerland
chron_data(country = "Switzerland", material = c("charcoal", "bone"))

#> # A tibble: 197 × 21
#>       id labnr    reference source_database    bp   std cal_bp cal_std delta_c13
#>    <int> <chr>    <lgl>     <chr>           <int> <int>  <int>   <int>     <dbl>
#>  1  7878 ETH-111… NA        calpal           3845    60   4261     164         0
#>  2  7880 ETH-261… NA        calpal           5050    60   5801     135         0
#>  3  7887 ETH-976… NA        calpal           4905    65   5643     188         0
#>  4  7888 ETH-976… NA        calpal           4910    65   5648     188         0
#>  5  7889 ETH-976… NA        calpal           4825    70   5540     190         0
#>  6  7890 ETH-976… NA        calpal           4965    60   5699     171         0
#>  7 12851 ETH-261… NA        radon            5050    60   5801     135        NA
#>  8 13975 CRC-1075 NA        radon            4245   165   4795     478         0
#>  9 14013 B-2113   NA        radon            5780   270   6610     613        NA
#> 10 16324 CRC-354  NA        radon            3700    85   4044     259         0
#> # … with 187 more rows, and 12 more variables: site <chr>, site_phase <chr>,
#> #   feature <chr>, period <lgl>, typochronological_unit <lgl>,
#> #   ecochronological_unit <lgl>, material <chr>, lat <chr>, lng <chr>,
#> #   country <chr>, species <chr>, site_type <chr>